requirements.txt
Flask==1.1.2
gast==0.2.2
grpcio==1.27.2
h5py==2.10.0
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
MarkupSafe==1.1.1
numpy==1.18.1
oauthlib==3.1.0
opt-einsum==3.1.0
Pillow==7.1.2
protobuf==3.11.4
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycparser==2.20
PyJWT==1.7.1
pyOpenSSL==19.1.0
pyreadline==2.1
PySocks==1.7.1
PyYAML==5.3.1
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
scipy==1.4.1
six==1.15.0
tensorboard==2.2.1
tensorboard-plugin-wit==1.6.0
tensorflow==2.1.0
tensorflow-estimator==2.1.0
termcolor==1.1.0
urllib3==1.25.8
Werkzeug==0.16.1
win-inet-pton==1.1.0
wincertstore==0.2
wrapt==1.12.1
